requirements.txt
# FastAPI and server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Pydantic with email support (newer versions have pre-built wheels)
pydantic>=2.10.0
pydantic[email]>=2.10.0

# MongoDB
pymongo>=4.10.0

# Environment variables
python-dotenv>=1.0.0

# Optional but recommended
python-multipart>=0.0.12
